Consolidate and fix download file functions
