+++ app/src/js/components/ReconciliationReports/gnf-report.js
@@ -1,29 +1,15 @@
-/* eslint-disable no-restricted-syntax */
-/* eslint-disable guard-for-in */
 import PropTypes from 'prop-types';
 import React from 'react';
 import {
   searchReconciliationReport,
   clearReconciliationSearch,
-  filterReconciliationReport,
-  clearReconciliationReportFilter,
 } from '../../actions';
 import SortableTable from '../SortableTable/SortableTable';
-import { reshapeReport } from './reshape-report';
-import { downloadFile } from '../../utils/download-file';
 import Search from '../Search/search';
-import Dropdown from '../DropDown/dropdown';
 import ReportHeading from './report-heading';
+import { handleDownloadJsonClick } from '../../utils/download-file';
 import { tableColumnsGnf } from '../../utils/table-config/reconciliation-reports';
 import { getCollectionId } from '../../utils/format';
-
-const bucketsForFilter = (allBuckets) => {
-  const uniqueBuckets = [...new Set(allBuckets)];
-  return uniqueBuckets.map((bucket) => ({
-    id: bucket,
-    label: bucket,
-  }));
-};
 
 const GnfReport = ({
   dispatch,
@@ -52,49 +38,21 @@ const GnfReport = ({
     location: 'cumulus'
   }));
 
-  const allGranules = [...cmrGranules, ...cumulusGranules];
+  let allGranules = [...cmrGranules, ...cumulusGranules];
 
-  console.log(allGranules);
+  if (filterString) {
+    allGranules = allGranules.filter((granule) => granule.granuleId.toLowerCase()
+      .includes(filterString.toLowerCase()));
+  }
+
+  const reportState = allGranules.length > 0 ? 'CONFLICT' : 'PASSED';
 
   const downloadOptions = [
     {
       label: 'JSON - Full Report',
-      onClick: handleDownloadJsonClick,
+      onClick: (e) => handleDownloadJsonClick(e, { data: recordData, reportName }),
     },
   ];
-
-  function convertToCSV(data, columns) {
-    const csvHeader = columns.map((column) => column.accessor).join(',');
-
-    const csvData = data
-      .map((item) => {
-        let line = '';
-        for (const prop in item) {
-          if (line !== '') line += ',';
-          line += item[prop];
-        }
-        return line;
-      })
-      .join('\r\n');
-    return `${csvHeader}\r\n${csvData}`;
-  }
-
-  function handleDownloadJsonClick(e) {
-    e.preventDefault();
-    const jsonHref = `data:text/json;charset=utf-8,${encodeURIComponent(
-      JSON.stringify(record.data)
-    )}`;
-    downloadFile(jsonHref, `${reconciliationReportName}.json`);
-  }
-
-  function handleDownloadCsvClick(e, table) {
-    e.preventDefault();
-    const { name, data: tableData, columns: tableColumns } = table;
-    const data = convertToCSV(tableData, tableColumns);
-    const csvData = new Blob([data], { type: 'text/csv' });
-    const url = window.URL.createObjectURL(csvData);
-    downloadFile(url, `${reportName}-${name}.csv`);
-  }
 
   return (
     <div className="page__component">
@@ -103,6 +61,7 @@ const GnfReport = ({
         endTime={reportEndTime}
         error={error}
         name={reportName}
+        reportState={reportState}
         startTime={reportStartTime}
         type='Granule Not Found'
       />
@@ -115,16 +74,6 @@ const GnfReport = ({
             label="Search"
             placeholder="Search"
           />
-          {/* <Dropdown
-            action={filterReconciliationReport}
-            clear={clearReconciliationReportFilter}
-            options={bucketsForFilter(allBuckets)}
-            paramKey="bucket"
-            label="Bucket"
-            inputProps={{
-              placeholder: 'All',
-            }}
-          /> */}
         </div>
 
         <SortableTable
